Created Adapter interface and moved adapter implementations into their respective database files as inner classes for encapsulation. Should be easier to extend the native databases now since they'll all call processSIMNative method in ProcessLanguages.
