+++ jyhton/build.xml
@@ -209,7 +209,8 @@ oracle.jar=C:/workspace/HEAD/for_development/bisdevsrv28/jboss/server/infra/lib/
             <pathelement path="${extlibs.dir}/slf4j-log4j12-1.6.4.jar"/>
             <pathelement path="${extlibs.dir}/sqljdbc4.jar"/>
             <pathelement path="${extlibs.dir}/stringtemplate-3.2.1.jar"/>
-            <pathelement path="${extlibs.dir}/wdb.jar"/>
+            <!-- <pathelement path="${extlibs.dir}/wdb.jar"/> -->
+            <pathelement path="${extlibs.dir}/wdb-newest.jar"/>
             <pathelement path="${extlibs.dir}/xercesImpl-2.10.0.jar"/>
             <pathelement path="${extlibs.dir}/xml-apis-1.4.01.jar"/>
             <pathelement path="${extlibs.dir}/batches.jar" />
@@ -723,7 +724,7 @@ The readme text for the next release will be like:
             <zipfileset src="${extlibs.dir}/slf4j-log4j12-1.6.4.jar"/>
             <zipfileset src="${extlibs.dir}/sqljdbc4.jar"/>
             <zipfileset src="${extlibs.dir}/stringtemplate-3.2.1.jar"/>
-            <zipfileset src="${extlibs.dir}/wdb.jar"/>
+            <!-- <zipfileset src="${extlibs.dir}/wdb.jar"/> -->
             <zipfileset src="${extlibs.dir}/xercesImpl-2.10.0.jar"/>
             <zipfileset src="${extlibs.dir}/xml-apis-1.4.01.jar"/>
             <zipfileset src="${extlibs.dir}/batches.jar" />
